+++ app/reactive-item-service/src/main/java/org/harvanir/pattern/clean/item/app/entrypoint/v1/controller/ItemQueryControllerV1.java
@@ -43,7 +43,8 @@ public class ItemQueryControllerV1 {
   }
 
   @GetMapping("/{id}/{delaySeconds}")
-  public Mono<ItemResponseRest> findByIdAndDelaySeconds(@PathVariable Long id, @PathVariable Integer delaySeconds) {
+  public Mono<ItemResponseRest> findByIdAndDelaySeconds(
+      @PathVariable Long id, @PathVariable Float delaySeconds) {
     return Mono.defer(
         () -> {
           ItemFindPresenterImpl presenter = new ItemFindPresenterImpl(beanMapper);
