+++ core/item-entity/src/main/java/org/harvanir/pattern/clean/item/core/entity/FindWithDelayRequest.java
@@ -14,5 +14,5 @@ public class FindWithDelayRequest {
 
   private Long id;
 
-  private Integer delaySeconds;
+  private Float delaySeconds;
 }
